backend/requirements.txt
fastapi
magnum
uvicorn
boto3
python-dotenv
